+++ freemind/modes/mindmapmode/MindMapController.java
@@ -1525,7 +1525,7 @@ public class MindMapController extends ControllerAdapter implements
      */
 
 	public void increaseFontSize(MindMapNode node, int increment) {
-		int newSize = Integer.valueOf(node.getFontSize()).intValue()
+		int newSize = Integer.valueOf(node.getFontSize())
 				+ increment;
 
 		if (newSize > 0) {
